PB-675 move mask types
